requirements.txt
flask>=3.0.0
flask-cors>=4.0.0

opencv-python-headless>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
Pillow>=10.0.0

easyocr>=1.7.0
pytesseract>=0.3.10


